# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Email Microservice API",
    "description": "API untuk Email Microservice yang mendukung pengiriman email, manajemen template, dan pelacakan email.",
    "version": "1.0.0",
    "contact": {
      "name": "Tim API Email Microservice",
      "email": "api-support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Development Server"
    },
    {
      "url": "https://api.example.com/api/v1",
      "description": "Production Server"
    }
  ],
...
  "paths": {
    "/emails/manual": {
      "post": {
        "tags": ["Email"],
        "summary": "Mengirim email kustom",
        "description": "Mengirim email dengan konten yang ditentukan oleh pengguna",
        "operationId": "sendManualEmail",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "description": "Payload pengiriman email",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualEmailRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Email diantrekan untuk pengiriman",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validasi gagal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Autentikasi gagal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit terlampaui",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/emails/{id}/status": {
      "get": {
        "tags": ["Email"],
        "summary": "Mendapatkan status email",
        "description": "Mendapatkan status pengiriman email",
        "operationId": "getEmailStatus",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID email",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Status email berhasil diambil",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Autentikasi gagal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Email tidak ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ManualEmailRequest": {
        "type": "object",
        "required": ["to", "subject", "body"],
        "properties": {
          "to": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "email"
            },
            "description": "Daftar penerima email"
          },
          "cc": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "email"
            },
            "description": "Daftar carbon copy"
          },
          "bcc": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "email"
            },
            "description": "Daftar blind carbon copy"
          },
          "subject": {
            "type": "string",
            "description": "Subjek email"
          },
          "body": {
            "type": "object",
            "required": ["text"],
            "properties": {
              "text": {
                "type": "string",
                "description": "Konten email dalam format teks biasa"
              },
              "html": {
                "type": "string",
                "description": "Konten email dalam format HTML"
              }
            }
          },
          "attachments": {
            "type": "array",
            "items": {
              "type": "object",
              "required": ["filename", "content", "contentType"],
              "properties": {
                "filename": {
                  "type": "string",
                  "description": "Nama file lampiran"
                },
                "content": {
                  "type": "string",
                  "description": "Konten file dalam format base64"
                },
                "contentType": {
                  "type": "string",
                  "description": "MIME type file"
                }
              }
            }
          },
          "priority": {
            "type": "string",
            "enum": ["high", "normal", "low"],
            "default": "normal",
            "description": "Prioritas email"
          },
          "headers": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Header email kustom"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Metadata untuk keperluan tracking"
          }
        }
      },
...
      "EmailResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["success"]
          },
          "traceID": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "email_id": {
                "type": "string",
                "description": "ID unik email"
              }
            }
          }
        }
      },
      "EmailStatusResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["success"]
          },
          "traceID": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "description": "ID email"
              },
              "status": {
                "type": "string",
                "enum": [
                  "pending",
                  "sending",
                  "delivered",
                  "failed",
                  "bounced",
                  "rejected",
                  "complained"
                ],
                "description": "Status email"
              },
              "sentAt": {
                "type": "string",
                "format": "date-time",
                "description": "Waktu pengiriman"
              },
              "deliveredAt": {
                "type": "string",
                "format": "date-time",
                "description": "Waktu pengiriman berhasil"
              },
              "failureReason": {
                "type": "string",
                "nullable": true,
                "description": "Alasan kegagalan jika gagal"
              },
              "attempts": {
                "type": "integer",
                "description": "Jumlah percobaan pengiriman"
              },
              "provider": {
                "type": "string",
                "description": "Provider email yang digunakan"
              },
              "messageId": {
                "type": "string",
                "description": "ID message dari provider"
              }
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["error"]
          },
          "traceID": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "errors": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "field": {
                      "type": "string",
                      "description": "Nama field yang error"
                    },
                    "message": {
                      "type": "string",
                      "description": "Deskripsi error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "AdminTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Token"
      }
    }
  }
}